+++ src/main/java/chess/controller/ChessController.java
@@ -4,13 +4,13 @@ import chess.domain.member.Member;
 import chess.domain.pieces.Color;
 import chess.domain.pieces.Piece;
 import chess.domain.position.Position;
-import chess.dto.responseDto.BoardDto;
-import chess.dto.responseDto.GameStatusDto;
-import chess.dto.requestDto.MoveDto;
-import chess.dto.requestDto.NewGameInfoDto;
-import chess.dto.responseDto.RoomDto;
-import chess.dto.responseDto.RoomsDto;
-import chess.dto.responseDto.StatusDto;
+import chess.dto.response.BoardDto;
+import chess.dto.response.GameStatusDto;
+import chess.dto.request.MoveDto;
+import chess.dto.request.NewGameInfoDto;
+import chess.dto.response.RoomDto;
+import chess.dto.response.RoomsDto;
+import chess.dto.response.StatusDto;
 import chess.entities.ChessGame;
 import chess.service.GameService;
 import java.util.ArrayList;
