+++ src/main/java/chess/dto/response/BoardDto.java
@@ -1,4 +1,4 @@
-package chess.dto.responseDto;
+package chess.dto.response;
 
 import chess.domain.member.Member;
 import chess.domain.pieces.Piece;
