+++ src/main/java/chess/dto/response/RoomsDto.java
@@ -1,4 +1,4 @@
-package chess.dto.responseDto;
+package chess.dto.response;
 
 import java.util.List;
