+++ src/main/java/chess/dto/response/StatusDto.java
@@ -1,4 +1,4 @@
-package chess.dto.responseDto;
+package chess.dto.response;
 
 import java.util.List;
 import java.util.Map;
